requirements.txt
discord.py==1.7.3
requests==2.31.0
python-dotenv==1.0.0
colorama==0.4.6
